Added SDL and OpenGL dependency to project.
As SDL 2 doesn't seem to widespread yet we ship it in our
tree and compile it statically into our binary. The version
in the tree is a recent trunk revision of the project modified
to be able to link statically in /MT(d) mode.

Added a dependency on OpenGL as well as a link dependency
against the corresponding libraries

Expanded the 3dchess dummy app to make some use of SDL and OpenGL

Again not tested on Linux.