Fix Linux build.
Apparently future::wait_for only started to work properly
in gcc 4.8 . As we want to use that make it a requirement.
I guess boost would have the needed classes to workaround but
it's fairly simple to get the gcc.

Added a shell script to run cmake with the right parameters for
selecting the correct compiler version similar to the windows
bat script.