Add Open Asset Import Library dependency as submodule.
Currently referencing the git head of assimp as they don't
tag releases. If any issues show up a stable version should
be chosen.

For now build cmd and viewer tool. This incurs a DX dependency
on windows but might be helpful when handling models. Can be
switched off if it causes trouble.

Adjusted .gitignore to not ignore the submodules.