Added very rudimentary GameOver and win detection.
For now assume either having taken the king of the
other player or eradicating all of its pieces ends
the game. The former is obviously not quite rules
compliant and we might have to think of something
different. The latter is meant for debugging
situations in which we don't want to place kings.