Add piece square table support to MaterialEvaluator.
PSQT as well as piece values now consistently taken from
http://chessprogramming.wikispaces.com/Simplified+evaluation+function .
Thanks to the PSQTs the opening of the AI is much more sane.
It now opens itself much more possibilities later in the game.
This slows down calculation though but that's not that big
of an issue right now. Lots of low hanging fruit in the
performance department once the time comes to take it.

Added various needed helpers and did a bit of refactoring.