Extend scoring to special case game over conditions.
Previously the AI was purely focused on collecting as many
points as possible instead of ending the game in any way it
could. Now the scoring special cases those situations to award
a MAX_SCORE for a win, a MIN_SCORE for a loss and 0 for a draw.
This should fix #14.